pyproject.toml
[tool.poetry]
name = "wordle-solver"
version = "0.1.0"
description = "A Python Wordle solver that utilizes information theory for intelligent word prediction"
authors = ["luiherch <hernandezchirlaque@gmail.com>"]
license = "CC-BY-NC"
readme = "README.md"
packages = [{include = "wordle_solver"}]

[tool.poetry.dependencies]
python = "^3.10"
customtkinter = "^5.1.3"
numpy = "^1.24.3"
Pillow = "^9.5.0"
scipy = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
